refactor: collapse PR lifecycle projection into pull-request-snapshot
